Title: 합과 평균의 차, Time: 121ms, Memory: 24MB, Status: Passed - CodeTree

+++ 240312/합과 평균의 차/sub-of-average-and-sum.py
@@ -3,7 +3,7 @@ a,b,c = map(int, input().split())
 total = a+b+c
 
 print(total)
-avg = total / 3
+avg = int(total / 3)
 
 print(avg)
 print(total - avg)
